refactor to allow multiple htmls to be used for forest
