+++ forest4/styles/animation.css
@@ -92,3 +92,13 @@
   }
   100% { right: 15px }
 }
+
+
+@keyframes scale-in {
+  0% {
+    transform: scale(0);
+  }
+  100% {
+    transform: scale(1);
+  }
+}
