+++ forest4/styles/main.css
@@ -264,13 +264,16 @@ button.disable {
   border: 2px solid white;
   background-color: var(--light-green);
   top: -20px;
-  left: 20px;
+  /* left: 20px; */
+  left: 15px;
   overflow: hidden;
 }
 
 .text.name {
   position: absolute;
   top: 50px;
+  /* animation: scale-in forwards ease 0.5s;
+  left: 15px; */
   animation: slide-in-from-left forwards ease 0.5s;
   width: 68px;
   text-align: center;
@@ -304,6 +307,7 @@ button.disable {
 
 .choice {
   position: absolute;
+  /* right: 15px; */
   right: 100vh;
   bottom: 20px;
   background-color: #392a20d8;
@@ -311,6 +315,7 @@ button.disable {
   flex-direction: column;
   justify-content: space-between;
   align-items: center;
+  /* animation: scale-in forwards ease 0.5s; */
   animation: slide-in-from-right forwards ease 0.5s;
 }
 
@@ -356,17 +361,19 @@ button.disable {
 .choice {
   position: absolute;
   right: 100vh;
+  /* right: 15px; */
   bottom: 20px;
   background-color: #392a20d8;
   display: flex;
   flex-direction: column;
   justify-content: space-between;
   align-items: center;
+  /* animation: scale-in forwards ease 0.5s; */
   animation: slide-in-from-right forwards ease 0.5s;
 }
 
 .face {
-  animation: slide-in-from-left forwards ease 0.5s;
+  /* animation: scale-in forwards ease 0.5s; */
   z-index: 5;
 }
