+++ forest4/test-floor.html
@@ -6,9 +6,9 @@
 
   <link rel="stylesheet" href="./styles/main.css">
   <link rel="stylesheet" href="./styles/animation.css">
-  <script type="module" src ="./scripts/area/app.js" charset="utf-8"></script>
+  <script type="module" src ="./scripts/area/app2.js" charset="utf-8"></script>
 
-  <title>forest4</title>
+  <title>testFloor</title>
 </head>
 <body>
   
@@ -30,15 +30,6 @@
   </div>
   <button class="touch-toggle green-btn">touch: OFF</button>
   <div class="transition-cover intro flex-center">
-    <div class="intro-box">
-      <div class="intro-img">
-        <div class="waves"></div>
-        <div class="waves"></div>
-        <div class="waves"></div>
-        <div class="bear-on-duck"></div>
-      </div>
-      <button class="start-button green-btn hidden">start</button>
-    </div>
     <div class="artwork"></div>
   </div>
